requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==4.1.0
celery==5.3.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1.4
django-cors-headers==3.13.0
django-crontab==0.7.1
django-debug-toolbar==4.1.0
django-filter==22.1
django-import-export==3.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
et-xmlfile==1.1.0
flower==2.0.1
gunicorn==20.1.0
humanize==4.8.0
idna==3.4
importlib-metadata==5.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.2
Markdown==3.4.1
MarkupPy==1.14
MarkupSafe==2.1.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
Pillow==9.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7
PyYAML==6.0.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tablib==3.5.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.11.0
